Move 3 turn.rs helpers into existing siblings (parent #680)
Move 3 standalone helpers from `turn.rs` into existing sibling
modules instead of creating new files. Small follow-on cleanup in
the meta-issue #680 reduction effort.

Migrated to `lifecycle.rs` (`pub(super)`):

* `extract_plan_constraints(contents) -> Vec<String>` — extracts
    bullet items under the `## Constraints` section of a plan
    markdown document. Naturally belongs alongside the other plan
    document helpers (`plan_is_substantive`, `plan_missing_sections`,
    `plan_section_body`, etc.) already in `lifecycle.rs`.

Migrated to `tester.rs` (`pub(super)`):

* `tester_approval_mode(yes_mode, stdin_is_terminal) -> ApprovalMode`
    — picks the `ApprovalMode` for a tester smoke run.
* `run_tester_llm_call(&client, model, session_id, &prompt)
    -> Result<String, TesterLlmError>` — invokes the tester sidecar
    LLM, emits `agent.tester.llm_call_{started,completed,failed}`
    events, rejects replies with tool_calls. Naturally belongs in
    `tester.rs` alongside the rest of the Tester Skill v1 pipeline
    (`detect`, `materialize_test`, `run_smoke`, etc.).

`turn.rs` adjustments:

* Removed the 3 fn definitions.
* Caller sites rewritten:
  - `extract_plan_constraints(...)` → `lifecycle::extract_plan_constraints(...)`
    (`lifecycle::` is already in scope via the existing import).
  - `tester_approval_mode(...)` → `tester::tester_approval_mode(...)`
    (`tester::` is already in scope via the existing `use super::tester;`).
  - `run_tester_llm_call(...)` → `tester::run_tester_llm_call(...)`.
* Test mod ref `super::tester_approval_mode` rewritten to
  `super::super::tester::tester_approval_mode`.

LOC delta:

* `turn.rs`: 27,987 → 27,905 (-82 LOC)
* `lifecycle.rs`: +28 LOC
* `tester.rs`: +75 LOC
* Cumulative `turn.rs`: 39,489 → 27,905 (-11,584 LOC, -29.3%)

`pub(super)` matching pre-extraction visibility. No new sibling
module created (DR3-001 maintained).

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
